+++ packages/logger/src/logger.ts
@@ -129,7 +129,7 @@ export class Logger {
    * @param message {any} message or Error object
    * @param args {...any=} arguments
    */
-  public error(message: string | Error, ...args: any[]): Logger {
+  public error(message: string | Error | unknown, ...args: any[]): Logger {
     this.log('error', message, ...args)
 
     return this
@@ -197,7 +197,11 @@ export class Logger {
     return this
   }
 
-  private log(level: Level, message: string | Error, ...args: any): Logger {
+  private log(
+    level: Level,
+    message: string | Error | unknown,
+    ...args: any
+  ): Logger {
     if (this.silent) return this
 
     if (LevelPriority[level] < LevelPriority[this.level]) return this